Overhauled table creation, changed user creation
Table Creation
-Table creation moved to a setup method, which is called only once from
the master process in bin/www.
-Separated creation super-query into 1 query for the multiple tables,
wrapped by a promise that resolves if it's successful or rejects
otherwise.
-Log message when creating each table, and when all tables created.

Coin Creation
-Default/universal coin created along with the tables now. Supposed to
always be at id 1, so updates whatever it finds there if it already
exists.

User Creation
-User creation moved to a MySQL transaction.
-Creates an entry in the database for the user with the
default/universal coin, with an amount of 0. Done so that new users will
always have an entry in the database. Done with the user creation
transaction.